Fixing bulk scattering with reschedule logging
Bulk scattering scheduled items with a =org-log-reschedule=
value of "time" will prompt for a note for the first item and scatter
it, but then will silently fail to reschedule the remaining items.

This patch fixes this by storing the state of =org-log-reschedule= right
before rescheduling, setting it to nil, and returning it to its original
value after rescheduling.

This is obviously extremely naive. Input is welcome.